@@ -13,6 +13,9 @@
   "files": [
     "dist"
   ],
+  "engines": {
+    "node": "^18"
+  },
   "scripts": {
     "build": "tsc",
     "format:check": "prettier --check '{lib,test}/**/*.{js,ts,json,yml}'",
@@ -29,8 +32,9 @@
   "author": "snyk.io",
   "license": "Apache-2.0",
   "devDependencies": {
+    "@types/debug": "^4.1.12",
     "@types/jest": "^29.5.12",
-    "@types/node": "11.13.8",
+    "@types/node": "^18.11.18",
     "@types/sinon": "^7.5.2",
     "@typescript-eslint/eslint-plugin": "^5.31.0",
     "@typescript-eslint/parser": "^5.31.0",
@@ -43,13 +47,13 @@
     "tap": "^12.0.1",
     "tap-only": "0.0.5",
     "ts-jest": "^29.1.2",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.7.4"
+    "ts-node": "^10.9.2",
+    "typescript": "^4.9.5"
   },
   "dependencies": {
     "@snyk/cli-interface": "2.11.3",
     "@snyk/dep-graph": "^1.23.1",
-    "debug": "^4.1.1",
+    "debug": "^4.3.4",
     "glob": "^7.1.6",
     "packageurl-js": "^1.0.0",
     "shescape": "1.6.1",
